feat: enable direct access to signup/signin form from respective buttons
