+++ src/hooks/useValidation.js
@@ -19,6 +19,13 @@ const useValidation = (credentials, userSignup, clubSignup) => {
     });
   }
 
+  if (credentials.password !== credentials.confirmPassword) {
+    errors.push({
+      error: true,
+      message: "Password and confirm password do not match",
+    });
+  }
+
   if (userSignup) {
     if (!credentials.firstname) {
       errors.push({ error: true, message: "Please enter your first name" });
