+++ src/pages/route.js
@@ -1,13 +1,13 @@
 export { default as Home } from "./Home";
 
+//Auth Routes
+export { default as AuthRegister } from "./Auth/AuthRegister";
+export { default as AuthLogin } from "./Auth/AuthLogin";
+
 // User Routes
-export { default as UserRegister } from "./Auth/User/UserRegister";
-export { default as UserLogin } from "./Auth/User/UserLogin";
 export { default as UserProfile } from "./User/UserProfile";
 
 // Club Routes
-export { default as ClubRegister } from "./Auth/Club/ClubsRegister";
-export { default as ClubLogin } from "./Auth/Club/ClubsLogin";
 export { default as ClubProfile } from "./Club/Profile/ClubProfile";
 export { default as Clubs } from "./Club/AllClubs/Clubs";
 export { default as Events } from "./Event/AllEvents/Events";
